refact: Merge utils
Added unique functions from scraper_utils into utils.py and deleted file. Refactored scraper.py accordingly.
