+++ utils/utils.py
@@ -1,4 +1,5 @@
 import os
+import re
 from datetime import datetime
 from csv import DictReader
 from configparser import ConfigParser
@@ -21,6 +22,21 @@ def create_directory_if_missing(file_path: str):
     """
     directory = os.path.dirname(file_path)
     os.makedirs(directory, exist_ok=True)
+
+
+def get_bike_type_from_desc(desc):
+    bike_types_list = [  # order matters for fork, frame, kid, girl, and bmx as qualifiers
+        'frame', 'frameset', 'fork', 'kid', 'girl', 'e-bike', 'bmx', 'city', 'commuter', 'comfort',
+        'cruiser', 'fat', 'triathlon', 'adventure', 'touring', 'urban',
+        'track', 'road', 'mountain', 'cyclocross', 'hybrid',
+        'gravel'
+    ]
+
+    for bike_type in bike_types_list:
+        if re.search(re.escape(bike_type), desc, re.IGNORECASE):
+            return bike_type
+
+    return None
 
 
 def get_fieldnames_from_file(filepath: str) -> list:
